+++ datanode/sqlstore/deposits.go
@@ -145,8 +145,12 @@ func (d *Deposits) getByPartyCursorPagination(ctx context.Context, party string,
 func getDepositsByPartyQuery(party string, dateRange entities.DateRange) (string, []interface{}) {
 	var args []interface{}
 
-	query := fmt.Sprintf(`select id, status, party_id, asset, amount, foreign_tx_hash, credited_timestamp, created_timestamp, tx_hash, vega_time
-		from deposits_current where party_id = %s`, nextBindVar(&args, entities.PartyID(party)))
+	query := `select id, status, party_id, asset, amount, foreign_tx_hash, credited_timestamp, created_timestamp, tx_hash, vega_time
+		from deposits_current`
+
+	if party != "" {
+		query = fmt.Sprintf(`%s where party_id = %s`, query, nextBindVar(&args, entities.PartyID(party)))
+	}
 
 	return filterDateRange(query, depositsFilterDateColumn, dateRange, args...)
 }