+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -
 
 ### 🐛 Fixes
+- [6273](https://github.com/vegaprotocol/vega/issues/6273) - Deposits return an empty list when no party is provided
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -